102 Introduces hook to get the config from the store and expose a getOverriddenValue function
