+++ packages/orchestrator-ui-components/src/components/WfoSubscription/index.ts
@@ -1,5 +1,7 @@
 export * from './utils';
 
+export * from './overrides';
+
 export * from './SubscriptionKeyValueBlock';
 export * from './WfoInSyncField';
 export * from './WfoProcessesTimeline';
