+++ packages/orchestrator-ui-components/src/components/WfoSubscription/overrides/index.ts
@@ -1,0 +1,1 @@
+export * from './useSubscriptionDetailValueOverride';
