+++ packages/orchestrator-ui-components/src/contexts/index.ts
@@ -3,4 +3,3 @@ export * from './ConfirmationDialogProvider';
 export * from './OrchestratorConfigContext';
 export * from './PolicyContext';
 export * from './TreeContext';
-export * from './ValueOverrideContext';
